Updated Weapon Wheel
Updated, Restructured and Optimized the weapon wheel to remove all expensive calls in update. Now works based on input and only if input is being recieved then it will push though. A further optimization for this is to replace input checks for event based input. But we will first see how the optimization works for this after we have completed phase 2