Reconstruction AI
Removed all expensive calls, and separated out logic for the AI into methods. I have found a call that looks for a bool that is not apart of the animator which I will fix tomorrow. I will just write up things I have done and to do lists within the document to move forwards easily tomorrow