Player Data Pull
Player data can now be pulled from server on login, this is then implemented into a player class. This is all pulled and then will not be destroyed on load. Therefor being usable on any scene. Which once logging off or finishing game will push that data back to the database.