Finally! The server is now forcing registery for the sessions.
Right so now there are 2 errors to fix up, and then we can move to making this occur when the lobby scene is created, or even about to be created. Once this is we can move onto the major P2P stuff, and then into starting the game and host switching.