Server Update
The server now handles the fact that a player has either hosted or joined a server. Accounting for that and reducing the amount of players remaining that can join it. The next step is to reduce the player count on the case that the player leaves the server so it adds to the player slots available