Forwarded the connection to the server connection
The fix that needs doing firstly is to make the message that logs the player out of the server. This will need to then apply the leaving and align for a new host. This is the issue currently but if i add a button that should not be too much of an issue