Added in the communication logic for ready up and more (Read Desc)(
Added the communication between Master Server and Client. After some research, it looks like its a good idea to keep this connection open and then have a secondary connection that allows for connection as host and game movement areas. The aim for this will be to proceed with ready up and control scene changes. Where as the link for the master server will be feeding information about how many spaces are left, if the game is started etc. I will begin creating the full game server setup after my break