Hit a wall here:
Another wall has been hit, we need communication with the master server. So I decided to attempt to create 2 Network Managers 1 for master server communication, the other for creating P2P hosting sessions. But only 1 Network Manager is allowed to be created and held in the same scene so this approach is going to have to change. I will have to intergrate this functionality within the MasterServerLinkConnector instead, since this is the main communication, we will have to disconnect from the master server while the game continues then switch back.. Somehow I will have to push this. This goes back to my first thought, to log all the players into the same master server to start with, allow the lobby to proceed, only once the game is launched with it deter to the P2P connection. Once this game is finished, it logs data collected by the players reconnects back to the master server to maybe show stats and pushes then across into the database. This seems like the way to handle this.