Lobby is not interpolating the players correctly. Commiting before I go into NetworkRoomManager
There is a huge issue with the servers ability to coordinate specific players into a lobby. However, after searching through the api I found something that is for RoomManagement. Which this already has a prebuild api to help for this hopefully helping the focus move towards that and fix this issue.